be/requirements.txt
asgiref==3.10.0
attrs==25.4.0
boto3==1.35.0
botocore==1.35.99
click==8.3.1
Django==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==9.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
wait-for-it==2.3.0
python-jose==3.3.0
python-keycloak==4.10.0
requests==2.32.3
